Step 2: JPA persistence layer, admin API, and email notifications
Replace the in-memory Property/Enquiry data with a real persistence
layer: JPA entities, Specification-based filtering (type/status/price
range), and repositories, running on H2 (file-based, zero-setup) for
local dev and PostgreSQL via env vars for production.

Property now matches the agreed schema (title, type, price, size/
sizeUnit, status AVAILABLE/SOLD, featured, imageUrls, createdAt), scoped
to PLOT/VILLA/FARMHOUSE. A CommandLineRunner seeds 3 sample properties
on first boot only (idempotent across restarts).

Add secured admin endpoints (HTTP Basic, single user via ADMIN_USERNAME/
ADMIN_PASSWORD env vars) for property CRUD and enquiry listing, kept
under /api/admin/** and separate from the public, unauthenticated
GET/enquiry-submission endpoints. Enquiries persist to the DB and
trigger an opt-in email notification (app.mail.enabled, off by default
so missing SMTP config never breaks lead capture) with a safe no-op/
log-and-continue fallback on send failure.

Testimonials/gallery/stats stay in-memory (not in the brief's data
model, no admin management needed for MVP).

Note: the Angular property pages still reference the old field names
(name/startingPrice/features) — this is expected mid-refactor per the
agreed build order and will be updated in Steps 3-4.

+++ backend/src/main/java/com/squarebuild/infratech/service/StatsService.java
@@ -13,6 +13,6 @@ public class StatsService {
     }
 
     public CompanyStats getStats() {
-        return new CompanyStats(73000, propertyService.count(), 480, 25, 2025);
+        return new CompanyStats(73000, (int) propertyService.count(), 480, 25, 2025);
     }
 }
